requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.3.2
cloudinary==1.41.0
dj-database-url==2.3.0
Django==4.2.19
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
googlemaps==4.10.0
gunicorn==23.0.0
idna==3.8
msgpack==1.1.0
multidict==6.1.0
packaging==24.2
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.10
PyJWT==2.9.0
pyotp==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
twilio==9.3.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yarl==1.11.1
